requirements.txt
gradio
Pillow
matplotlib
numpy
cryptography
torch
scipy